pyproject.toml
[project]
name = "drifter"
version = "0.1.0"
description = ""
authors = [
    {name = "Cameron Reid",email = "csreid.cr@gmail.com"}
]
requires-python = ">=3.12,<3.13"
dependencies = [
    "pybullet (>=3.2.7,<4.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "xacro (>=1.13.3,<2.0.0)",
    "noise (>=1.2.2,<2.0.0)",
    "pygame (>=2.6.1,<3.0.0)",
    "ruff (>=0.11.12,<0.12.0)",
    "scipy (>=1.15.3,<2.0.0)",
    "gymnasium (>=1.2.0,<2.0.0)",
    "stable-baselines3 (>=2.7.0,<3.0.0)",
    "tensorboard>=2.20.0,<3.0.0",
    "tqdm (>=4.67.1,<5.0.0)",
    "rich (>=14.1.0,<15.0.0)",
    "sb3-contrib (>=2.7.0,<3.0.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "pyyaml>=6.0.2",
    "tabulate>=0.9.0",
    "ipython>=9.6.0",
    "jupyterlab>=4.4.9",
    "pytorch-mppi>=0.8.0",
    "types-tqdm>=4.67.0.20250809",
    "types-tabulate>=0.9.0.20241207",
    "pandas-stubs>=2.3.2.250926",
    "mypy>=1.18.2",
]

[tool.ruff]
line-length=80

[tool.ruff.format]
indent-style="tab"

[tool.uv]
package = false

[tool.mypy]
[[tool.mypy.overrides]]
module = "pybullet.*"
ignore_missing_imports = true
[[tool.mypy.overrides]]
module = "noise.*"
ignore_missing_imports = true
[[tool.mypy.overrides]]
module = "scipy.*"
ignore_missing_imports = true
[[tool.mypy.overrides]]
module = "scipy.*"
ignore_missing_imports = true
[[tool.mypy.overrides]]
module = [
    "pybullet.*",
    "pybullet_data.*",  # Add these
    "xacro.*",
    "cv2.*",
    "pygame.*",
    "gymnasium.*",
    "stable_baselines3.*",
    "scipy.*",
    "IPython.*",
    "pytorch_mppi.*",  # Add this
]
ignore_missing_imports = true

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
